% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAllFiles.R
\name{checkAllFiles}
\alias{checkAllFiles}
\title{Title function to check if files in user-input directory are suitable for fluidigmValidationMain or passSummary}
\usage{
checkAllFiles(dirFiles, funInd)
}
\arguments{
\item{dirFiles}{: directory of files}

\item{funInd}{: 1 for fluidigmValidationMain and 2 for passSummary}
}
\description{
Title function to check if files in user-input directory are suitable for fluidigmValidationMain or passSummary
}
\author{
Johnson Liu
}

