% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFC.R
\name{checkFC}
\alias{checkFC}
\title{Title function to check FC}
\usage{
checkFC(data)
}
\arguments{
\item{data}{: data with INFO column splited}
}
\value{
data that met FC criteria that rows with only Silent_something, Silent, Synonymous_something, Synonymous are deleted
}
\description{
Title function to check FC
}

