% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPass.R
\name{filterPASS}
\alias{filterPASS}
\title{Title function to only acquire data rows whose FILTER column is a PASS}
\usage{
filterPASS(df)
}
\arguments{
\item{df}{: a data frame of data}
}
\value{
part of input data frame whose FILTER column is a PASS
}
\description{
Title function to only acquire data rows whose FILTER column is a PASS
}

