% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPassDepth.R
\name{filterPassDepth}
\alias{filterPassDepth}
\title{Title function to subset data that has passed filter and minimum depth requiremtn}
\usage{
filterPassDepth(data, minDepth)
}
\arguments{
\item{data}{: original data, filter with PASS and other measures and depths of all values}

\item{minDepth}{: minimum depth requirement}
}
\value{
data that has passed filter and at or deeper than the minimum depth
}
\description{
Title function to subset data that has passed filter and minimum depth requiremtn
}

