% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterQualityFrequency.R
\name{filterQualityFrequency}
\alias{filterQualityFrequency}
\title{Title function to filter out data that do not meet minimum qualiy and frequency requirement}
\usage{
filterQualityFrequency(df)
}
\arguments{
\item{df}{: a data frame of data.}
}
\value{
part of input data frame that meet minimum quality and frequency requirements.
}
\description{
Title function to filter out data that do not meet minimum qualiy and frequency requirement
}

