% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidigmValidationMain.R
\name{fluidigmValidationMain}
\alias{fluidigmValidationMain}
\title{Title Main function to perform analysis of Fluidigm data sets}
\usage{
fluidigmValidationMain()
}
\value{
an excel file of data labeled, filtered, and only cases with repetitive pos values
}
\description{
Title Main function to perform analysis of Fluidigm data sets
}

