% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatientID.R
\name{getPatientID}
\alias{getPatientID}
\title{Title function to obtain patient ID from file name}
\usage{
getPatientID(fileName)
}
\arguments{
\item{fileName}{: name of file}
}
\value{
The patient ID name
}
\description{
Title function to obtain patient ID from file name
}
\examples{
"PAT-ID33-DNA-ID141-Curettage-E4_S122" returns "ID33"
}

