% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelMutationGroup.R
\name{labelMutationGroup}
\alias{labelMutationGroup}
\title{Title function for labelling mutation group}
\usage{
labelMutationGroup(data)
}
\arguments{
\item{data}{: data with tumor type (i.e. T1, T2, N) labelled}
}
\value{
data with mutation group labelled (i.e. somatic, artifact, germline, normal) and only whose positions are repetitive
}
\description{
Title function for labelling mutation group
}
\author{
Johnson Liu
}

