% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelNT.R
\name{labelNT}
\alias{labelNT}
\title{Title function to label genes according to file name in a data frame of merged data files. If file name contains NORMAL, genes are labeled "N", otherwise,t hey are labeled "T" for tumor}
\usage{
labelNT(df)
}
\arguments{
\item{df}{: a data frame of all data merged frome several xlsx files}
}
\value{
original data frame with labels
}
\description{
Title function to label genes according to file name in a data frame of merged data files. If file name contains NORMAL, genes are labeled "N", otherwise,t hey are labeled "T" for tumor
}

