% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelTumorType.R
\name{labelTumorType}
\alias{labelTumorType}
\title{Title function to label tumor type based on file name}
\usage{
labelTumorType(data)
}
\arguments{
\item{data}{: fluidigm data}
}
\value{
data whose rows are labeled "T1" if coming from a file with "Curettage", "T2" if coming
from a file containing "OP" and "N" if from a file containing "NORMAL"
}
\description{
Title function to label tumor type based on file name
}

