% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFile.R
\name{mergeFile}
\alias{mergeFile}
\title{Title function to merge all the xlsx files in a user-input directory}
\usage{
mergeFile()
}
\value{
a data frame of all data merged from the files in the user-defined directory
}
\description{
Title function to merge all the xlsx files in a user-input directory
}

