% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moreFiles.R
\name{moreFiles}
\alias{moreFiles}
\title{Title function for creating a file for each patient}
\usage{
moreFiles(allRepetitiveData)
}
\arguments{
\item{allRepetitiveData}{: a data frame with only data with repetitive positions}
}
\value{
write an excel file for each patient
}
\description{
Title function for creating a file for each patient
}

