% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notMutationGroupException.R
\name{notMutationGroupException}
\alias{notMutationGroupException}
\title{Title function to check if a chunk is an mutation group exception (e.g have more than 1 N, or more than 2 T1 or T2)}
\usage{
notMutationGroupException(vec)
}
\arguments{
\item{vec}{: input vector for tumor type}
}
\value{
TRUE is it is exception false otherwise
}
\description{
Title function to check if a chunk is an mutation group exception (e.g have more than 1 N, or more than 2 T1 or T2)
}
\author{
Johnson Liu
}

