% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onceN.R
\name{onceN}
\alias{onceN}
\title{Title function to check if a vector only has 1 "N"}
\usage{
onceN(vec)
}
\arguments{
\item{vec}{: vector to be checked}
}
\value{
true if the input vector has only 1 "N" and false otherwise
}
\description{
Title function to check if a vector only has 1 "N"
}
\author{
Johnson Liu
}

