% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onceT1.R
\name{onceT1}
\alias{onceT1}
\title{Title function to check if a vector has only 1 T1}
\usage{
onceT1(vec)
}
\arguments{
\item{vec}{: input vector to be checked}
}
\value{
true if vec has only 1 T1 and false otherwise
}
\description{
Title function to check if a vector has only 1 T1
}

