% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onceT2.R
\name{onceT2}
\alias{onceT2}
\title{Title function to check if a vector has only 1 T2}
\usage{
onceT2(vec)
}
\arguments{
\item{vec}{: input vector to be checked}
}
\value{
true if vec has only 1 T2 and false otherwise
}
\description{
Title function to check if a vector has only 1 T2
}

