% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyRepetitivePos.R
\name{onlyRepetitivePos}
\alias{onlyRepetitivePos}
\title{Title function to find interesting cases (cases with repetitive positions)}
\usage{
onlyRepetitivePos(data)
}
\arguments{
\item{data}{}
}
\value{
data of only interesting cases
}
\description{
Title function to find interesting cases (cases with repetitive positions)
}

