% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{organize}
\alias{organize}
\title{Title function to create a folder for each patient and move all files of that patient to that folder created}
\usage{
organize(dirFiles)
}
\arguments{
\item{dirFiles}{: directory of all original patient files}
}
\description{
Title function to create a folder for each patient and move all files of that patient to that folder created
}

