% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAmpliconInfo.R
\name{processAmpliconInfo}
\alias{processAmpliconInfo}
\title{Title function to perform amplicon information}
\usage{
processAmpliconInfo(ampliconInfo, data)
}
\arguments{
\item{ampliconInfo}{: excel sheet for amplicon info, titled "Mutation for every amplicons.xls"}

\item{data}{: data from filterPassDepth with depth setting 0}
}
\value{
data with depth for each amplicon from each file
}
\description{
Title function to perform amplicon information
}

