% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameToTxt.R
\name{renameToTxt}
\alias{renameToTxt}
\title{Title function to rename vcf files to txt files in not using Windows machine if the files are not already txt files}
\usage{
renameToTxt(dirFiles)
}
\arguments{
\item{dirFiles}{: directory for all files at the beginning of the program}
}
\description{
Title function to rename vcf files to txt files in not using Windows machine if the files are not already txt files
}
\examples{
"PAT-ID95-DNA-ID149-OP-D2_S12.vcf" would become "PAT-ID95-DNA-ID149-OP-D2_S12.txt"

}

