% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitINFO.R
\name{splitINFO}
\alias{splitINFO}
\title{Title function to parse the column INFO by semicolon}
\usage{
splitINFO(df)
}
\arguments{
\item{df}{: data frame of all data}
}
\value{
a data frame whose INFO column has been split into 5 columns: DP, TI, GI, FC, EXON
}
\description{
Title function to parse the column INFO by semicolon
}

