% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitNUMS.R
\name{splitNUMS}
\alias{splitNUMS}
\title{Title function to split NUMS column by colon}
\usage{
splitNUMS(df)
}
\arguments{
\item{df}{: a data frame}
}
\value{
original data frame with NUMS column split into : "lala","Quality","AllelicRatio","frequency","Unkown1","Unknown2","Unknown3"
}
\description{
Title function to split NUMS column by colon
}

