% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_AxB.R
\name{SS_AxB}
\alias{SS_AxB}
\title{Title function to calculate interaction between two groups (SS_AxB) in a two way annova test}
\usage{
SS_AxB(ms)
}
\arguments{
\item{ms}{: list for all matrices}
}
\value{
SS_AxB
}
\description{
Title function to calculate interaction between two groups (SS_AxB) in a two way annova test
}

