% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twiceT1.R
\name{twiceT1}
\alias{twiceT1}
\title{Title function to check if a vector has only 2 T1}
\usage{
twiceT1(vec)
}
\arguments{
\item{vec}{: input vector to be checked}
}
\value{
true if vec has only 2 T1 and false otherwise
}
\description{
Title function to check if a vector has only 2 T1
}
\author{
Johnson Liu
}

