% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twiceT2.R
\name{twiceT2}
\alias{twiceT2}
\title{Title function to check if a vector has only 2 T2}
\usage{
twiceT2(vec)
}
\arguments{
\item{vec}{: input vector to be checked}
}
\value{
true if vec has only 2 T2 and false otherwise
}
\description{
Title function to check if a vector has only 2 T2
}

