% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterQualityFrequency.R
\name{filterQualityFrequency}
\alias{filterQualityFrequency}
\title{Title function to filter quality and frequency for fluidigmValidationMain}
\usage{
filterQualityFrequency(data, minQual, minFreq)
}
\arguments{
\item{data}{: data with columns FILTER and NUMS from original fluidigm machine fiels}

\item{minQual}{: minimum quality number entered by user}

\item{minFreq}{: minimum frequency entered by user}
}
\value{
modified data with filter only equal to PASS and those that meet minimum frequency and quality requirements.
}
\description{
Title function to filter quality and frequency for fluidigmValidationMain
}
\author{
Johnson Liu
}

