% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidigmValidationMain.R
\name{fluidigmValidationMain}
\alias{fluidigmValidationMain}
\title{Title main function for processing fluidigm data files}
\usage{
fluidigmValidationMain()
}
\description{
Title main function for processing fluidigm data files
}

