% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyRepetitivePos.R
\name{onlyRepetitivePos}
\alias{onlyRepetitivePos}
\title{Title function for finding only data rows whose positions are repetitive}
\usage{
onlyRepetitivePos(data)
}
\arguments{
\item{data}{whose positions are some repetitive and some not}
}
\value{
part of input data with repetitive positions
}
\description{
Title function for finding only data rows whose positions are repetitive
}

