% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitINFO.R
\name{splitINFO}
\alias{splitINFO}
\title{Title function to split INFO column in fluidigm data}
\usage{
splitINFO(data)
}
\arguments{
\item{data}{with INFO column not splited}
}
\value{
data with INFO columns splited into DP, TI, GI, FC and only data with EXON
}
\description{
Title function to split INFO column in fluidigm data
}

